refactor: wheelcolumn으로 UI 컴포넌트 분리
